requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
ase=3.21.1=pypi_0
blas=1.0=mkl
boto3=1.18.21=pyhd3eb1b0_0
botocore=1.21.41=pyhd3eb1b0_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pypi_0
cornac=1.14.2=py38h56ea1f7_0
cryptography=35.0.0=py38hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cute-ranking=0.0.3=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
deeprobust=0.2.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.0=pypi_0
freetype=2.10.4=h5ab3b9f_0
gensim=3.8.3=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
googledrivedownloader=0.4=pypi_0
h5py=3.2.1=pypi_0
huggingface-hub=0.2.1=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
intel-openmp=2021.2.0=h06a4308_610
isodate=0.6.0=pypi_0
jinja2=3.0.1=pypi_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=11.2.0=h69a702a_14
libgfortran5=11.2.0=h5c6108e_14
libgomp=11.2.0=h1d223b6_14
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
liblapack=3.9.0=9_mkl
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.36.0=pypi_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=pypi_0
matplotlib-base=3.4.3=py38hf4fb855_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
mpmath=1.2.1=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.5.1=pypi_0
ninja=1.10.2=hff7bd54_1
numba=0.53.1=pypi_0
numpy=1.19.4=pypi_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f98852_0
packaging=21.3=pypi_0
pandas=1.2.4=pypi_0
pillow=8.2.0=py38he98fc37_0
pip=21.1.2=py38h06a4308_0
powerlaw=1.4.6=pyh9f0ad1d_1
protobuf=3.17.3=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygraphviz=1.7=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.1=pypi_0
python-decouple=3.4=pypi_0
python-louvain=0.15=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-ignite=0.4.8=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0=pypi_0
rdflib=5.0.0=pypi_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.25.1=pypi_0
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.46=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.3=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
smart-open=5.1.0=pypi_0
spark-parser=1.8.9=pypi_0
sqlite=3.35.4=hdfb4753_0
tensorboardx=2.2=pypi_0
texttable=1.6.3=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.6.6=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
torch-cluster=1.5.9=pypi_0
torch-geometric=1.7.0=pypi_0
torch-scatter=2.0.7=pypi_0
torch-sparse=0.6.9=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu102
tornado=6.1=py38h497a2fe_2
tqdm=4.61.0=pypi_0
transformers=4.12.5=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
unpyclib=0.8.1=pypi_0
urllib3=1.26.5=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xdis=6.0.3=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
